Direct use of private constants instead of getters.

+++ L2J_Mobius_2.5_Underground/java/com/l2jmobius/gameserver/network/clientpackets/alchemy/RequestAlchemyTryMixCube.java
@@ -120,7 +120,7 @@ public class RequestAlchemyTryMixCube implements IClientIncomingPacket
 				return;
 			}
 			
-			final int price = itemInstance.getReferencePrice();
+			final long price = itemInstance.getReferencePrice();
 			if (itemInstance.getReferencePrice() == 0)
 			{
 				player.sendPacket(SystemMessageId.THIS_ITEM_CANNOT_BE_COMBINED);
